initial implementation of `kallithea_version` fact
The implementation of the `kallithea_version` fact has been heavily
inspired by the elasticsearch/elasticsearch module.

The idea is to get the Kallithea version from the output of `pip version
kallithea`. However, the problem is that the path to the pip binary of
Kallitheas virtualenv depends on `$kallithea::install::venv`, which is
unknown on the puppet client. Therefore, we first store this information
on the Puppet client in `/etc/kallithea.yml` which enables us to read
them later, when the `kallithea_version` fact needs to be evaluated.

The concept of persisting interesting information on the client side
seems to be so useful, that I have generalized this by means of the
`settings_store` class and the `stored_setting` defined type. They are
based on the richardc/datacat module, which allows us to merge the data
given by multiple `stored_setting` declarations before they get written
to `/etc/kallithea.yml`.

refs #10